fix double-precision on soft-float target, negative-modulo rotation, jelly-frame clamp
Latent correctness/perf issues found while working through the performance
pass:

- effects-utils.cpp, geometry.h, geometry.cpp, utils.cpp: several float
  computations used bare double literals (255.0, 1000.0, 180.0) or the
  double overloads of sqrt()/sin(), silently promoting otherwise-float math
  to double. Cold path everywhere except geometry's rotation, but the C3 has
  no FPU at all, so every one of these pays a software double-precision
  routine instead of the hardware/soft-float-optimized single-precision one.
  One-character fixes (f suffix, sqrtf/sinf) throughout.

- Geometry::applyGlobalRandomRotation(): cdegrees - tcDeg is negative for
  roughly half the LEDs at any given rotation angle, and C++'s % truncates
  toward zero rather than floors - so "% FULL_CIRCLE" left the result
  negative there, which wrapped to 36000..65535 once stored in the uint16_t
  cdegrees field. Masked today because no shipping effect with the
  ROTATE_SPACE hint reads raw angles, but ClockSweep's getCoordinate()
  already does. Add FULL_CIRCLE before the modulo. Pinned with a test that
  applies the rotation at 20 different random angles and asserts every LED's
  cdegrees stays in [0, FULL_CIRCLE).

- JellyFrame::colorIndex(): cast baseHue_ + displacement*gain to uint8_t
  without clamping first - StandingWaveRing's equivalent already clamps.
  float->uint8_t on an out-of-range value is UB, not a wrap. Clamp before
  the cast, matching StandingWaveRing.

+++ include/geometry/geometry.h
@@ -35,7 +35,7 @@ struct PolarCoordinates
         // Calculate radius
         float x_sq = (float)cart.x * cart.x;
         float y_sq = (float)cart.y * cart.y;
-        radius = (uint16_t)sqrt(x_sq + y_sq);
+        radius = (uint16_t)sqrtf(x_sq + y_sq);
 
         // Calculate angle in centi-degrees (0-36000 for 0-360 degrees)
         float angle_rad = atan2f((float)cart.y, (float)cart.x);
